Avoid deprecated usage of ModuleTemplate
