+++ lib/EvalDevToolModulePlugin.js
@@ -5,15 +5,23 @@
 
 "use strict";
 
-const EvalDevToolModuleTemplatePlugin = require("./EvalDevToolModuleTemplatePlugin");
+const { RawSource } = require("webpack-sources");
+const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
+const ModuleFilenameHelpers = require("./ModuleFilenameHelpers");
 
+/** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("./Compiler")} Compiler */
+
+/** @type {WeakMap<Source, Source>} */
+const cache = new WeakMap();
 
 class EvalDevToolModulePlugin {
 	constructor(options) {
-		this.sourceUrlComment = options.sourceUrlComment;
-		this.moduleFilenameTemplate = options.moduleFilenameTemplate;
-		this.namespace = options.namespace;
+		this.namespace = options.namespace || "";
+		this.sourceUrlComment = options.sourceUrlComment || "\n//# sourceURL=[url]";
+		this.moduleFilenameTemplate =
+			options.moduleFilenameTemplate ||
+			"webpack://[namespace]/[resourcePath]?[loaders]";
 	}
 
 	/**
@@ -22,12 +30,46 @@ class EvalDevToolModulePlugin {
 	 */
 	apply(compiler) {
 		compiler.hooks.compilation.tap("EvalDevToolModulePlugin", compilation => {
-			new EvalDevToolModuleTemplatePlugin({
-				compilation,
-				sourceUrlComment: this.sourceUrlComment,
-				moduleFilenameTemplate: this.moduleFilenameTemplate,
-				namespace: this.namespace
-			}).apply(compilation.moduleTemplates.javascript);
+			const hooks = JavascriptModulesPlugin.getCompilationHooks(compilation);
+			hooks.renderModuleContent.tap(
+				"EvalDevToolModulePlugin",
+				(source, module, { runtimeTemplate, chunkGraph }) => {
+					const cacheEntry = cache.get(source);
+					if (cacheEntry !== undefined) return cacheEntry;
+					const content = source.source();
+					const str = ModuleFilenameHelpers.createFilename(
+						module,
+						{
+							moduleFilenameTemplate: this.moduleFilenameTemplate,
+							namespace: this.namespace
+						},
+						{
+							requestShortener: runtimeTemplate.requestShortener,
+							chunkGraph
+						}
+					);
+					const footer =
+						"\n" +
+						this.sourceUrlComment.replace(
+							/\[url\]/g,
+							encodeURI(str)
+								.replace(/%2F/g, "/")
+								.replace(/%20/g, "_")
+								.replace(/%5E/g, "^")
+								.replace(/%5C/g, "\\")
+								.replace(/^\//, "")
+						);
+					const result = new RawSource(
+						`eval(${JSON.stringify(content + footer)});`
+					);
+					cache.set(source, result);
+					return result;
+				}
+			);
+			hooks.chunkHash.tap("EvalDevToolModulePlugin", (chunk, hash) => {
+				hash.update("EvalDevToolModulePlugin");
+				hash.update("1");
+			});
 		});
 	}
 }
